refactor: Route entity 및 repository 클래스 위치 변경
