+++ hub-service/src/main/java/org/sixpang/hubservice/domain/model/entity/Route.java
@@ -1,4 +1,4 @@
-package org.sixpang.routeservice.domain.model;
+package org.sixpang.hubservice.domain.model.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
