+++ hub-service/src/main/java/org/sixpang/hubservice/domain/repository/RouteRepository.java
@@ -1,4 +1,4 @@
-package org.sixpang.routeservice.domain.repository;
+package org.sixpang.hubservice.domain.repository;
 
 import org.sixpang.routeservice.domain.model.Route;
