+++ hub-service/src/main/java/org/sixpang/hubservice/infrastructure/repository/RouteJpaRepository.java
@@ -1,4 +1,4 @@
-package org.sixpang.routeservice.infrastructure.repository;
+package org.sixpang.hubservice.infrastructure.repository;
 
 import org.sixpang.routeservice.domain.model.Route;
 import org.springframework.data.jpa.repository.JpaRepository;
