+++ hub-service/src/main/java/org/sixpang/hubservice/infrastructure/repository/RouteRepositoryImpl.java
@@ -1,4 +1,4 @@
-package org.sixpang.routeservice.infrastructure.repository;
+package org.sixpang.hubservice.infrastructure.repository;
 
 import lombok.RequiredArgsConstructor;
 import org.sixpang.routeservice.domain.model.Route;
